Refactor operation fixtures
