+++ p2panda-rs/src/test_utils/fixtures/new_fixtures/mod.rs
@@ -1,3 +1,5 @@
 mod entry_fixtures;
+mod operation_fixtures;
 
 pub use entry_fixtures::*;
+pub use operation_fixtures::*;
