+++ p2panda-rs/src/test_utils/fixtures/params.rs
@@ -10,18 +10,13 @@ use rand::Rng;
 use rstest::fixture;
 
 use crate::document::{DocumentId, DocumentViewId};
-use crate::entry::EntrySigned;
 use crate::hash::Hash;
 use crate::identity::{Author, KeyPair};
-use crate::operation::{
-    Operation, OperationEncoded, OperationFields, OperationId, OperationValue, OperationWithMeta,
-};
+use crate::operation::OperationId;
 use crate::schema::SchemaId;
 use crate::test_utils::constants::{DEFAULT_HASH, DEFAULT_PRIVATE_KEY, TEST_SCHEMA_ID};
-use crate::test_utils::fixtures::defaults;
-use crate::test_utils::utils;
-
 use crate::test_utils::fixtures::*;
+use crate::test_utils::utils;
 
 /// Fixture which injects the default private key string into a test method.
 #[fixture]
@@ -81,13 +76,6 @@ pub fn document_view_id(#[default(vec![DEFAULT_HASH])] hash_str_vec: Vec<&str>) 
     DocumentViewId::new(&hashes).unwrap()
 }
 
-/// Fixture which injects the default `OperationId` into a test method. Default value can be
-/// overridden at testing time by passing in a custom hash string.
-#[fixture]
-pub fn operation_id(#[default(DEFAULT_HASH)] hash_str: &str) -> OperationId {
-    OperationId::new(hash(hash_str))
-}
-
 /// Fixture which injects a random hash into a test method.
 #[fixture]
 pub fn random_hash() -> Hash {
@@ -95,62 +83,10 @@ pub fn random_hash() -> Hash {
     Hash::new_from_bytes(random_data).unwrap()
 }
 
-/// Fixture which injects a random operation id into a test method.
-#[fixture]
-pub fn random_operation_id() -> OperationId {
-    random_hash().into()
-}
-
 /// Fixture which injects a random document id.
 #[fixture]
 pub fn random_document_id() -> DocumentId {
     DocumentId::new(random_hash().into())
-}
-
-/// Fixture which injects a random document view id into a test method.
-#[fixture]
-pub fn random_document_view_id() -> DocumentViewId {
-    DocumentViewId::new(&[random_hash().into(), random_hash().into()]).unwrap()
-}
-
-/// Fixture which injects the default OperationFields value into a test method.
-///
-/// Default value can be overridden at testing time by passing in a custom vector of key-value
-/// tuples.
-#[fixture]
-pub fn fields(
-    #[default(vec![("message", defaults::operation_value())])] fields_vec: Vec<(
-        &str,
-        OperationValue,
-    )>,
-) -> OperationFields {
-    utils::operation_fields(fields_vec)
-}
-
-/// Fixture which injects the default OperationFields value into a test method.
-///
-/// Default value can be overridden at testing time by passing in a custom vector of key-value
-/// tuples.
-#[fixture]
-pub fn some_fields(
-    #[default(vec![("message", defaults::operation_value())])] fields_vec: Vec<(
-        &str,
-        OperationValue,
-    )>,
-) -> Option<OperationFields> {
-    Some(utils::operation_fields(fields_vec))
-}
-
-/// Fixture which injects the default Operation into a test method.
-///
-/// Default value can be overridden at testing time by passing in custom operation fields and
-/// document id.
-#[fixture]
-pub fn operation(
-    #[from(some_fields)] fields: Option<OperationFields>,
-    #[default(None)] previous_operations: Option<DocumentViewId>,
-) -> Operation {
-    utils::any_operation(fields, previous_operations)
 }
 
 /// Fixture which injects the default Hash into a test method as an Option.
@@ -160,91 +96,4 @@ pub fn operation(
 pub fn some_hash(#[default(DEFAULT_HASH)] str: &str) -> Option<Hash> {
     let hash = Hash::new(str);
     Some(hash.unwrap())
-}
-
-#[fixture]
-pub fn operation_encoded(operation: Operation) -> OperationEncoded {
-    OperationEncoded::try_from(&operation).unwrap()
-}
-
-/// Fixture which injects the default CREATE Operation into a test method.
-///
-/// Default value can be overridden at testing time by passing in custom schema hash and operation
-/// fields.
-#[fixture]
-pub fn create_operation(schema: SchemaId, fields: OperationFields) -> Operation {
-    utils::create_operation(schema, fields)
-}
-
-/// Fixture which injects the default UPDATE Operation into a test method.
-///
-/// Default value can be overridden at testing time by passing in custom schema hash, document id
-/// hash and operation fields.
-#[fixture]
-pub fn update_operation(
-    schema: SchemaId,
-    #[default(document_view_id(vec![DEFAULT_HASH]))] previous_operations: DocumentViewId,
-    #[default(fields(vec![("message", OperationValue::Text("Updated, hello!".to_string()))]))]
-    fields: OperationFields,
-) -> Operation {
-    utils::update_operation(schema, previous_operations, fields)
-}
-
-/// Fixture which injects the default DELETE Operation into a test method.
-///
-/// Default value can be overridden at testing time by passing in custom schema hash and document
-/// id hash.
-#[fixture]
-pub fn delete_operation(
-    schema: SchemaId,
-    #[default(document_view_id(vec![DEFAULT_HASH]))] previous_operations: DocumentViewId,
-) -> Operation {
-    utils::delete_operation(schema, previous_operations)
-}
-
-/// Fixture which injects a CREATE `OperationWithMeta` into a test method.
-#[fixture]
-pub fn create_operation_with_meta(
-    create_operation: Operation,
-    public_key: Author,
-    #[from(random_operation_id)] operation_id: OperationId,
-) -> OperationWithMeta {
-    OperationWithMeta::new_test_operation(&operation_id, &public_key, &create_operation)
-}
-
-/// Fixture which injects an UPDATE OperationWithMeta into a test method.
-#[fixture]
-pub fn update_operation_with_meta(
-    update_operation: Operation,
-    public_key: Author,
-    #[from(random_operation_id)] operation_id: OperationId,
-) -> OperationWithMeta {
-    OperationWithMeta::new_test_operation(&operation_id, &public_key, &update_operation)
-}
-
-/// Fixture which injects a DELETE `OperationWithMeta` into a test method.
-#[fixture]
-pub fn delete_operation_with_meta(
-    delete_operation: Operation,
-    public_key: Author,
-    #[from(random_operation_id)] operation_id: OperationId,
-) -> OperationWithMeta {
-    OperationWithMeta::new_test_operation(&operation_id, &public_key, &delete_operation)
-}
-
-#[fixture]
-pub fn encoded_create_string(create_operation: Operation) -> String {
-    OperationEncoded::try_from(&create_operation)
-        .unwrap()
-        .as_str()
-        .to_owned()
-}
-
-/// Fixture which injects the default CREATE `OperationWithMeta` into a test method.
-#[fixture]
-pub fn meta_operation(
-    entry_signed_encoded: EntrySigned,
-    operation_encoded: OperationEncoded,
-) -> OperationWithMeta {
-    utils::meta_operation(entry_signed_encoded, operation_encoded)
 }
