+++ p2panda-rs/src/test_utils/utils.rs
@@ -9,15 +9,10 @@
 //! into `rstest` defined methods.
 use serde::Serialize;
 
-use crate::document::DocumentViewId;
-use crate::entry::{Entry, EntrySigned, LogId, SeqNum};
+use crate::entry::{LogId, SeqNum};
 use crate::hash::Hash;
 use crate::identity::KeyPair;
-use crate::operation::{
-    Operation, OperationEncoded, OperationFields, OperationValue, OperationWithMeta,
-};
 use crate::schema::SchemaId;
-use crate::test_utils::constants::TEST_SCHEMA_ID;
 
 /// A custom `Result` type to be able to dynamically propagate `Error` types.
 pub type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
@@ -38,40 +33,6 @@ pub struct NextEntryArgs {
     pub log_id: LogId,
 }
 
-/// A helper method for easily generating an operation of any type (CREATE, UPDATE or DELETE).
-///
-/// If a value for `fields` is provided, this is a CREATE operation. If values for both `fields`
-/// and `document_id` are provided, this is an UPDATE operation. If no value for `fields` is
-/// provided, this is a DELETE operation.
-pub fn any_operation(
-    fields: Option<OperationFields>,
-    previous_operations: Option<DocumentViewId>,
-) -> Operation {
-    let schema_id = SchemaId::new(TEST_SCHEMA_ID).unwrap();
-    match fields {
-        // It's a CREATE operation
-        Some(fields) if previous_operations.is_none() => {
-            Operation::new_create(schema_id, fields).unwrap()
-        }
-        // It's an UPDATE operation
-        Some(fields) => {
-            Operation::new_update(schema_id, previous_operations.unwrap(), fields).unwrap()
-        }
-        // It's a DELETE operation
-        None => Operation::new_delete(schema_id, previous_operations.unwrap()).unwrap(),
-    }
-}
-
-/// Helper method for generating OperationFields from a vector of key-value tuples, currently only
-/// string types are implemented.
-pub fn operation_fields(fields: Vec<(&str, OperationValue)>) -> OperationFields {
-    let mut operation_fields = OperationFields::new();
-    for (key, value) in fields.iter() {
-        operation_fields.add(key, value.to_owned()).unwrap();
-    }
-    operation_fields
-}
-
 /// Generate a new key pair, not based on the default private key.
 pub fn new_key_pair() -> KeyPair {
     KeyPair::new()
@@ -90,31 +51,4 @@ pub fn hash(hash_str: &str) -> Hash {
 /// Generate an application schema based on a schema id string.
 pub fn schema(schema_id: &str) -> SchemaId {
     SchemaId::new(schema_id).unwrap()
-}
-
-/// Generate a CREATE operation based on passed schema id and operation fields.
-pub fn create_operation(schema: SchemaId, fields: OperationFields) -> Operation {
-    Operation::new_create(schema, fields).unwrap()
-}
-
-/// Generate an UPDATE operation based on passed schema id, document id and operation fields.
-pub fn update_operation(
-    schema: SchemaId,
-    previous_operations: DocumentViewId,
-    fields: OperationFields,
-) -> Operation {
-    Operation::new_update(schema, previous_operations, fields).unwrap()
-}
-
-/// Generate a DELETE operation based on passed schema id and document id.
-pub fn delete_operation(schema: SchemaId, previous_operations: DocumentViewId) -> Operation {
-    Operation::new_delete(schema, previous_operations).unwrap()
-}
-
-/// Generate a CREATE meta-operation based on passed encoded entry and operation.
-pub fn meta_operation(
-    entry_signed_encoded: EntrySigned,
-    operation_encoded: OperationEncoded,
-) -> OperationWithMeta {
-    OperationWithMeta::new_from_entry(&entry_signed_encoded, &operation_encoded).unwrap()
 }
